requirements.txt
asgiref==3.2.10
config==0.5.0.post0
croniter==0.3.35
Django==3.1.2
django-scheduler==0.9.3
icalendar==4.0.7
natsort==7.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
schedule==0.6.0
six==1.15.0
sqlparse==0.4.1
